Removed singletons for command line options and replaced them with variables now included in lsst/ctrl/orca/__init__.py, and cleaned up code that used to use the singletons.
In DatabaseConfigurator.py:

Moved the database authorization lookup into the superclass, where it belonged.

Changed the database credential policy format slightly to accomodate multiple
databases, and ports.

Added new behavior for looking up the host, and how it matches against the
credentials policy file, as follows:

In the policy for databases

1) If only "database.host" is specified, it will match against the first
"database.authinfo.host" that matches it in the credentials policy, regardless
of the "database.authinfo.port".

2) If "database.host" and "database.port" are specified, it will match
against the first "database.authinfo.host" and "database.authinfo.port"
in the credentials policy.

3) If there is no match, an exception is thrown.